requirements.txt
aiogram==2.4
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
multidict==6.0.5
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pyparsing==3.1.2
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
